refactor: extract openrouter-response.ts from openrouter.ts (#261)
Move response parsing, validation, and stream processing utilities into a
dedicated module. The OpenRouterProvider class remains in openrouter.ts
with SDK construction and Sentry-instrumented generation.
No behavioral changes — all 17 provider tests pass unchanged.

Changes:
- Extract openrouter-response.ts (~285 lines): types (TextPart, StreamDeltaLike, StreamChoiceLike, StreamEventLike), type guards, parseContent, extractChunkText, normalizeUsage, validateNonStreamingResponse, streamFromEvents, getStatusCodeFromError
- Reduce openrouter.ts to ~342 lines: OpenRouterProvider class + SDK + Sentry spans
- Update imports within openrouter.ts to reference new module
