+++ src/features/ai/providers/openrouter.ts
@@ -7,9 +7,16 @@ import {
   ProviderInvalidResponseError,
 } from '@/features/ai/providers/errors';
 import {
-  asyncIterableToReadableStream,
-  toStream,
-} from '@/features/ai/streaming/utils';
+  type StreamEventLike,
+  isAsyncIterable,
+  isObjectRecord,
+  streamFromEvents,
+  validateNonStreamingResponse,
+  parseContent,
+  normalizeUsage,
+  getStatusCodeFromError,
+} from '@/features/ai/providers/openrouter-response';
+import { toStream } from '@/features/ai/streaming/utils';
 import {
   DEFAULT_GENERATION_EXTENSION_MS,
   DEFAULT_GENERATION_TIMEOUT_MS,
@@ -40,284 +47,6 @@ export type OpenRouterProviderConfig = {
 
 const OPENROUTER_DEFAULT_TIMEOUT_MS = DEFAULT_GENERATION_TIMEOUT_MS;
 const OPENROUTER_TIMEOUT_EXTENSION_MS = DEFAULT_GENERATION_EXTENSION_MS;
-
-type TextPart = {
-  type: string;
-  text?: string;
-};
-
-type StreamDeltaLike = {
-  content?: string | TextPart[] | null;
-};
-
-type StreamChoiceLike = {
-  delta?: StreamDeltaLike | null;
-  message?: StreamDeltaLike | null;
-};
-
-type StreamEventLike = {
-  choices?: StreamChoiceLike[];
-  usage?: {
-    promptTokens?: number;
-    completionTokens?: number;
-    totalTokens?: number;
-    input_tokens?: number;
-    output_tokens?: number;
-    total_tokens?: number;
-  };
-  delta?: string;
-};
-
-function isObjectRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null;
-}
-
-function isAsyncIterable(
-  value: unknown
-): value is AsyncIterable<StreamEventLike> {
-  return (
-    isObjectRecord(value) &&
-    Symbol.asyncIterator in value &&
-    typeof value[Symbol.asyncIterator] === 'function'
-  );
-}
-
-function parseContent(
-  content: string | TextPart[] | null | undefined
-): string | null {
-  if (!content) {
-    return null;
-  }
-
-  if (typeof content === 'string') {
-    return content;
-  }
-
-  const text = content
-    .filter((item): item is TextPart => item.type === 'text')
-    .map((item) => item.text ?? '')
-    .join('');
-
-  return text.length > 0 ? text : null;
-}
-
-function extractChunkText(event: StreamEventLike): string {
-  if (typeof event.delta === 'string' && event.delta.length > 0) {
-    return event.delta;
-  }
-
-  const choice = event.choices?.[0];
-  if (!choice) {
-    return '';
-  }
-
-  const fromDelta = parseContent(choice.delta?.content);
-  if (fromDelta) {
-    return fromDelta;
-  }
-
-  const fromMessage = parseContent(choice.message?.content);
-  return fromMessage ?? '';
-}
-
-function normalizeUsage(
-  usage:
-    | {
-        promptTokens?: number;
-        completionTokens?: number;
-        totalTokens?: number;
-        input_tokens?: number;
-        output_tokens?: number;
-        total_tokens?: number;
-      }
-    | undefined
-): ProviderUsage {
-  return {
-    promptTokens: usage?.promptTokens ?? usage?.input_tokens,
-    completionTokens: usage?.completionTokens ?? usage?.output_tokens,
-    totalTokens: usage?.totalTokens ?? usage?.total_tokens,
-  };
-}
-
-const USAGE_TOKEN_FIELDS = [
-  'promptTokens',
-  'completionTokens',
-  'totalTokens',
-  'input_tokens',
-  'output_tokens',
-  'total_tokens',
-] as const;
-
-function isTextPartArray(value: unknown): value is TextPart[] {
-  return (
-    Array.isArray(value) &&
-    value.every((part) => {
-      if (!isObjectRecord(part) || typeof part.type !== 'string') {
-        return false;
-      }
-      return part.text === undefined || typeof part.text === 'string';
-    })
-  );
-}
-
-function isUsageShape(value: unknown): value is StreamEventLike['usage'] {
-  if (!isObjectRecord(value)) {
-    return false;
-  }
-
-  return USAGE_TOKEN_FIELDS.every((field) => {
-    const fieldValue = value[field];
-    return fieldValue === undefined || typeof fieldValue === 'number';
-  });
-}
-
-function describeResponseValue(value: unknown): string {
-  if (value === null) {
-    return 'null';
-  }
-  if (value === undefined) {
-    return 'undefined';
-  }
-  if (Array.isArray(value)) {
-    return `array(length=${value.length})`;
-  }
-  if (isObjectRecord(value)) {
-    const keys = Object.keys(value);
-    return `object(keys=${keys.length > 0 ? keys.join(', ') : 'none'})`;
-  }
-  if (typeof value === 'string') {
-    return `string(length=${value.length})`;
-  }
-  if (
-    typeof value === 'number' ||
-    typeof value === 'boolean' ||
-    typeof value === 'bigint'
-  ) {
-    return `${typeof value}(${value.toString()})`;
-  }
-  if (typeof value === 'symbol') {
-    return 'symbol';
-  }
-  if (typeof value === 'function') {
-    return 'function';
-  }
-  return typeof value;
-}
-
-function createInvalidShapeError(
-  fieldPath: string,
-  expected: string,
-  actual: unknown
-): ProviderInvalidResponseError {
-  return new ProviderInvalidResponseError(
-    `OpenRouter returned invalid response shape: expected ${fieldPath} to be ${expected}, received ${describeResponseValue(actual)}`
-  );
-}
-
-function validateNonStreamingResponse(response: unknown): {
-  rawContent: string | TextPart[];
-  usage: StreamEventLike['usage'] | undefined;
-} {
-  if (!isObjectRecord(response)) {
-    throw createInvalidShapeError('response', 'an object', response);
-  }
-
-  const rawChoices = response.choices;
-  if (!Array.isArray(rawChoices)) {
-    throw createInvalidShapeError('choices', 'an array', rawChoices);
-  }
-  const choices: unknown[] = rawChoices;
-
-  if (choices.length === 0) {
-    throw new ProviderInvalidResponseError(
-      'OpenRouter returned an empty response (choices array was empty)'
-    );
-  }
-
-  const firstChoice = choices[0];
-  if (!isObjectRecord(firstChoice)) {
-    throw createInvalidShapeError('choices[0]', 'an object', firstChoice);
-  }
-
-  const message = firstChoice.message;
-  if (!isObjectRecord(message)) {
-    throw createInvalidShapeError('choices[0].message', 'an object', message);
-  }
-
-  const rawContent = message.content;
-  if (typeof rawContent !== 'string' && !isTextPartArray(rawContent)) {
-    throw createInvalidShapeError(
-      'choices[0].message.content',
-      'a string or TextPart[]',
-      rawContent
-    );
-  }
-
-  const usage = response.usage;
-  if (usage !== undefined && !isUsageShape(usage)) {
-    throw createInvalidShapeError(
-      'usage',
-      'an object with numeric token fields',
-      usage
-    );
-  }
-
-  return {
-    rawContent,
-    usage,
-  };
-}
-
-function getStatusCodeFromError(error: unknown): number | undefined {
-  if (!isObjectRecord(error)) {
-    return undefined;
-  }
-
-  if (typeof error.status === 'number') {
-    return error.status;
-  }
-
-  if (typeof error.statusCode === 'number') {
-    return error.statusCode;
-  }
-
-  if (
-    isObjectRecord(error.response) &&
-    typeof error.response.status === 'number'
-  ) {
-    return error.response.status;
-  }
-
-  return undefined;
-}
-
-function streamFromEvents(params: {
-  events: AsyncIterable<StreamEventLike>;
-  onUsage: (usage: ProviderUsage) => void;
-}): ReadableStream<string> {
-  const { events, onUsage } = params;
-  const textChunks = (async function* () {
-    let emittedAnyText = false;
-
-    for await (const event of events) {
-      onUsage(normalizeUsage(event.usage));
-      const text = extractChunkText(event);
-      if (!text) {
-        continue;
-      }
-      emittedAnyText = true;
-      yield text;
-    }
-
-    if (!emittedAnyText) {
-      throw new ProviderInvalidResponseError(
-        'OpenRouter returned no text content'
-      );
-    }
-  })();
-
-  return asyncIterableToReadableStream(textChunks);
-}
 
 export class OpenRouterProvider implements AiPlanGenerationProvider {
   private readonly client: OpenRouterClient;
